Move nested type to partial part file
